+++ src/gameplayer.h
@@ -94,16 +94,16 @@ private:
   uint32_t m_TotalPacketsReceived;          // the total number of packets received from the player
   uint32_t m_LeftCode;                      // the code to be sent in W3GS_PLAYERLEAVE_OTHERS for why this player left the game
   uint32_t m_SyncCounter;                   // the number of keepalive packets received from this player
-  uint32_t m_JoinTime;                      // GetTime when the player joined the game (used to delay sending the /whois a few seconds to allow for some lag)
+  uint64_t m_JoinTime;                      // GetTime when the player joined the game (used to delay sending the /whois a few seconds to allow for some lag)
   uint32_t m_LastMapPartSent;               // the last mappart sent to the player (for sending more than one part at a time)
   uint32_t m_LastMapPartAcked;              // the last mappart acknowledged by the player
-  uint32_t m_StartedDownloadingTicks;       // GetTicks when the player started downloading the map
-  uint32_t m_FinishedDownloadingTime;       // GetTime when the player finished downloading the map
-  uint32_t m_FinishedLoadingTicks;          // GetTicks when the player finished loading the game
-  uint32_t m_StartedLaggingTicks;           // GetTicks when the player started laggin
-  uint32_t m_LastGProxyWaitNoticeSentTime;  // GetTime when the last disconnection notice has been sent when using GProxy++
+  uint64_t m_StartedDownloadingTicks;       // GetTicks when the player started downloading the map
+  uint64_t m_FinishedDownloadingTime;       // GetTime when the player finished downloading the map
+  uint64_t m_FinishedLoadingTicks;          // GetTicks when the player finished loading the game
+  uint64_t m_StartedLaggingTicks;           // GetTicks when the player started laggin
+  uint64_t m_LastGProxyWaitNoticeSentTime;  // GetTime when the last disconnection notice has been sent when using GProxy++
   uint32_t m_GProxyReconnectKey;            // the GProxy++ reconnect key
-  uint32_t m_LastGProxyAckTime;             // GetTime when we last acknowledged GProxy++ packet
+  uint64_t m_LastGProxyAckTime;             // GetTime when we last acknowledged GProxy++ packet
   uint8_t m_PID;                            // the player's PID
   bool m_Spoofed;                           // if the player has spoof checked or not
   bool m_Reserved;                          // if the player is reserved (VIP) or not
@@ -144,14 +144,14 @@ public:
   inline std::string GetJoinedRealm() const                           { return m_JoinedRealm; }
   inline uint32_t GetLeftCode() const                                 { return m_LeftCode; }
   inline uint32_t GetSyncCounter() const                              { return m_SyncCounter; }
-  inline uint32_t GetJoinTime() const                                 { return m_JoinTime; }
+  inline uint64_t GetJoinTime() const                                 { return m_JoinTime; }
   inline uint32_t GetLastMapPartSent() const                          { return m_LastMapPartSent; }
   inline uint32_t GetLastMapPartAcked() const                         { return m_LastMapPartAcked; }
-  inline uint32_t GetStartedDownloadingTicks() const                  { return m_StartedDownloadingTicks; }
-  inline uint32_t GetFinishedDownloadingTime() const                  { return m_FinishedDownloadingTime; }
-  inline uint32_t GetFinishedLoadingTicks() const                     { return m_FinishedLoadingTicks; }
-  inline uint32_t GetStartedLaggingTicks() const                      { return m_StartedLaggingTicks; }
-  inline uint32_t GetLastGProxyWaitNoticeSentTime() const             { return m_LastGProxyWaitNoticeSentTime; }
+  inline uint64_t GetStartedDownloadingTicks() const                  { return m_StartedDownloadingTicks; }
+  inline uint64_t GetFinishedDownloadingTime() const                  { return m_FinishedDownloadingTime; }
+  inline uint64_t GetFinishedLoadingTicks() const                     { return m_FinishedLoadingTicks; }
+  inline uint64_t GetStartedLaggingTicks() const                      { return m_StartedLaggingTicks; }
+  inline uint64_t GetLastGProxyWaitNoticeSentTime() const             { return m_LastGProxyWaitNoticeSentTime; }
   inline uint32_t GetGProxyReconnectKey() const                       { return m_GProxyReconnectKey; }
   inline bool GetGProxy() const                                       { return m_GProxy; }
   inline bool GetGProxyDisconnectNoticeSent() const                   { return m_GProxyDisconnectNoticeSent; }
@@ -177,9 +177,9 @@ public:
   inline void SetSyncCounter(uint32_t nSyncCounter)                                    { m_SyncCounter = nSyncCounter; }
   inline void SetLastMapPartSent(uint32_t nLastMapPartSent)                            { m_LastMapPartSent = nLastMapPartSent; }
   inline void SetLastMapPartAcked(uint32_t nLastMapPartAcked)                          { m_LastMapPartAcked = nLastMapPartAcked; }
-  inline void SetStartedDownloadingTicks(uint32_t nStartedDownloadingTicks)            { m_StartedDownloadingTicks = nStartedDownloadingTicks; }
-  inline void SetFinishedDownloadingTime(uint32_t nFinishedDownloadingTime)            { m_FinishedDownloadingTime = nFinishedDownloadingTime; }
-  inline void SetStartedLaggingTicks(uint32_t nStartedLaggingTicks)                    { m_StartedLaggingTicks = nStartedLaggingTicks; }
+  inline void SetStartedDownloadingTicks(uint64_t nStartedDownloadingTicks)            { m_StartedDownloadingTicks = nStartedDownloadingTicks; }
+  inline void SetFinishedDownloadingTime(uint64_t nFinishedDownloadingTime)            { m_FinishedDownloadingTime = nFinishedDownloadingTime; }
+  inline void SetStartedLaggingTicks(uint64_t nStartedLaggingTicks)                    { m_StartedLaggingTicks = nStartedLaggingTicks; }
   inline void SetSpoofed(bool nSpoofed)                                                { m_Spoofed = nSpoofed; }
   inline void SetReserved(bool nReserved)                                              { m_Reserved = nReserved; }
   inline void SetWhoisShouldBeSent(bool nWhoisShouldBeSent)                            { m_WhoisShouldBeSent = nWhoisShouldBeSent; }
@@ -192,7 +192,7 @@ public:
   inline void SetMuted(bool nMuted)                                                    { m_Muted = nMuted; }
   inline void SetLeftMessageSent(bool nLeftMessageSent)                                { m_LeftMessageSent = nLeftMessageSent; }
   inline void SetGProxyDisconnectNoticeSent(bool nGProxyDisconnectNoticeSent)          { m_GProxyDisconnectNoticeSent = nGProxyDisconnectNoticeSent; }
-  inline void SetLastGProxyWaitNoticeSentTime(uint32_t nLastGProxyWaitNoticeSentTime)  { m_LastGProxyWaitNoticeSentTime = nLastGProxyWaitNoticeSentTime; }
+  inline void SetLastGProxyWaitNoticeSentTime(uint64_t nLastGProxyWaitNoticeSentTime)  { m_LastGProxyWaitNoticeSentTime = nLastGProxyWaitNoticeSentTime; }
 
   // processing functions
